requirements.txt
streamlit>=1.28.0
torch>=2.6.0
torchvision>=0.17.0
opencv-python>=4.8.0
mediapipe>=0.10.5
deepface>=0.0.75
tensorflow>=2.15.0
tf-keras>=2.15.0
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
